requirements-mobile.txt
lava-dl==0.10.2.rc1
torch==2.4.0
numpy
pqcrypto-dilithium-kyber
